Added new post and started modifying styles
I'm back! OK so basically I added one of the posts that I plan to add.
Or at least a draft of it. Still reads pretty much like crap. I'm not
getting across what I want to get across so I gotta work on that.

I also modified a lot of styles, mostly color and layout (which
technically is all of it :P).